Rename network package to io to reflect broader scope
