+++ app/src/main/java/com/hacktx/android/activities/CheckInActivity.java
@@ -44,7 +44,7 @@ import com.google.zxing.WriterException;
 import com.google.zxing.common.BitMatrix;
 import com.google.zxing.qrcode.QRCodeWriter;
 import com.hacktx.android.R;
-import com.hacktx.android.network.UserStateStore;
+import com.hacktx.android.io.UserStateStore;
 import com.hacktx.android.utils.HackTXUtils;
 
 import java.io.File;
