+++ app/src/main/java/com/hacktx/android/activities/EventDetailActivity.java
@@ -49,9 +49,9 @@ import com.hacktx.android.R;
 import com.hacktx.android.models.EventFeedback;
 import com.hacktx.android.models.ScheduleEvent;
 import com.hacktx.android.models.ScheduleSpeaker;
-import com.hacktx.android.network.HackTxClient;
-import com.hacktx.android.network.UserStateStore;
-import com.hacktx.android.network.services.HackTxService;
+import com.hacktx.android.io.HackTxClient;
+import com.hacktx.android.io.UserStateStore;
+import com.hacktx.android.io.services.HackTxService;
 import com.hacktx.android.utils.AlphaSatColorMatrixEvaluator;
 import com.hacktx.android.utils.ConfigParam;
 import com.hacktx.android.views.CircularImageView;
