+++ app/src/main/java/com/hacktx/android/fragments/AnnouncementFragment.java
@@ -33,8 +33,8 @@ import android.widget.TextView;
 
 import com.hacktx.android.R;
 import com.hacktx.android.models.Announcement;
-import com.hacktx.android.network.HackTxClient;
-import com.hacktx.android.network.services.HackTxService;
+import com.hacktx.android.io.HackTxClient;
+import com.hacktx.android.io.services.HackTxService;
 import com.hacktx.android.views.SpacesItemDecoration;
 import com.hacktx.android.views.adapters.AnnouncementsRecyclerView;
