+++ app/src/main/java/com/hacktx/android/io/UserStateStore.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hacktx.android.network;
+package com.hacktx.android.io;
 
 import android.content.Context;
 import android.content.SharedPreferences;
