+++ app/src/main/java/com/hacktx/android/io/services/HackTxService.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hacktx.android.network.services;
+package com.hacktx.android.io.services;
 
 import java.util.ArrayList;
